requirements.txt
aiosqlite==0.17.0
anyio==3.5.0
asgiref==3.5.0
click==8.1.3
colorama==0.4.4
fastapi==0.75.2
h11==0.13.0
httptools==0.4.0
idna==3.3
iso8601==1.0.2
pydantic==1.9.0
pypika-tortoise==0.1.4
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
sniffio==1.2.0
starlette==0.17.1
tortoise-orm==0.19.0
typing-extensions==4.2.0
uvicorn==0.17.6
watchgod==0.8.2
websockets==10.3
